+++ src/main/java/com/opuscapita/peppol/outbound/errors/OxalisError.java
@@ -1,0 +1,32 @@
+package com.opuscapita.peppol.outbound.errors;
+
+public class OxalisError {
+
+    private String mask;
+    private OutboundError type;
+
+    public OxalisError() {
+    }
+
+    public OxalisError(OutboundError type, String mask) {
+        this.type = type;
+        this.mask = mask;
+    }
+
+    public OutboundError getType() {
+        return type;
+    }
+
+    public void setType(OutboundError type) {
+        this.type = type;
+    }
+
+    public String getMask() {
+        return mask;
+    }
+
+    public void setMask(String mask) {
+        this.mask = mask;
+    }
+
+}
