+++ src/main/resources/application-local.properties
@@ -10,7 +10,7 @@ peppol.storage.blob.port=23012
 peppol.auth.url=http://localhost
 peppol.auth.port=23005
 
-spring.rabbitmq.host=http://localhost
+spring.rabbitmq.host=localhost
 spring.rabbitmq.port=35672
 spring.rabbitmq.username=${RABBITMQ_USER}
 spring.rabbitmq.password=${RABBITMQ_PASS}
