+++ src/main/resources/application.properties
@@ -2,4 +2,4 @@
 spring.application.name=peppol-outbound
 server.port=3040
 peppol.outbound.queue.in.name=peppol.message.outbound
-spring.profiles.include=commons,queue,snc
+spring.profiles.include=commons,errors,queue,snc
